Move Rename Analysis dialog to presenter
